move tree stuff to signle file
