+++ vscode/src/activateDeviceScript.ts
@@ -3,15 +3,9 @@
  *--------------------------------------------------------*/
 
 import {
-    DOCS_ROOT,
     Flags,
     FRAME_PROCESS,
-    identifierToUrlPath,
-    JDDevice,
     JDFrameBuffer,
-    JDRegister,
-    prettyUnit,
-    REGISTER_NODE_NAME,
     serializeToTrace,
 } from "jacdac-ts"
 import * as vscode from "vscode"
@@ -33,13 +27,10 @@ import {
 } from "./jacdac"
 import {
     JDomDeviceTreeItem,
-    JDomDeviceTreeDataProvider,
-    JDomTreeItem,
-    JDomWatchTreeDataProvider,
-    JDomRegisterTreeItem,
+    activateTreeViews,
 } from "./JDomTreeDataProvider"
 import { activateMainStatusBar } from "./mainstatusbar"
-import { DeviceScriptExtensionState, NodeWatch } from "./state"
+import { DeviceScriptExtensionState } from "./state"
 
 export function activateDeviceScript(context: vscode.ExtensionContext) {
     const { subscriptions, extensionMode, extension } = context
@@ -165,83 +156,11 @@ export function activateDeviceScript(context: vscode.ExtensionContext) {
         console.info = console.log
     }
 
-    subscriptions.push(
-        vscode.commands.registerCommand(
-            "extension.devicescript.node.copy",
-            async (item: JDomTreeItem) => {
-                if (!item) return
-                const { node } = item
-                const { nodeKind } = node
-                switch (nodeKind) {
-                    case REGISTER_NODE_NAME: {
-                        const reg = node as JDRegister
-                        const value = reg.humanValue
-                        await vscode.env.clipboard.writeText(
-                            `${reg.qualifiedName}: ${value}`
-                        )
-                        break
-                    }
-                }
-            }
-        )
-    )
-    subscriptions.push(
-        vscode.commands.registerCommand(
-            "extension.devicescript.node.select",
-            (item: JDomTreeItem) => {
-                if (!item) return
-                const { node } = item
-                const { nodeKind } = node
-                switch (nodeKind) {
-                    case REGISTER_NODE_NAME: {
-                        ;(node as JDRegister).scheduleRefresh()
-                        break
-                    }
-                }
-            }
-        )
-    )
-    const selectNodeCommand: vscode.Command = {
-        title: "select node",
-        command: "extension.devicescript.node.select",
-    }
-    const jdomTreeDataProvider = new JDomDeviceTreeDataProvider(
-        extensionState,
-        selectNodeCommand
-    )
-    vscode.window.registerTreeDataProvider(
-        "extension.devicescript.jdom-explorer",
-        jdomTreeDataProvider
-    )
-    const jdomWatchTreeDataProvider = new JDomWatchTreeDataProvider(
-        extensionState,
-        selectNodeCommand
-    )
-    vscode.window.registerTreeDataProvider(
-        "extension.devicescript.watch",
-        jdomWatchTreeDataProvider
-    )
     const cloudState = new CloudExtensionState(context, extensionState)
     registerCloudTreeDataProvider(cloudState)
     registerCloudStatusBar(cloudState)
 
     subscriptions.push(
-        vscode.commands.registerCommand(
-            "extension.devicescript.device.showFirmwareInformation",
-            (device: JDDevice) => {
-                if (!device) return
-                const spec =
-                    bus.deviceCatalog.specificationFromProductIdentifier(
-                        device.productIdentifier
-                    )
-                if (spec) {
-                    const uri = `${DOCS_ROOT}${`devices/${identifierToUrlPath(
-                        spec.id
-                    )}`}`
-                    vscode.env.openExternal(vscode.Uri.parse(uri))
-                }
-            }
-        ),
         vscode.commands.registerCommand(
             "extension.devicescript.simulator.stop",
             async () => {
@@ -257,162 +176,10 @@ export function activateDeviceScript(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand(
             "extension.devicescript.pickDeviceScriptManager",
             () => extensionState.pickDeviceScriptManager()
-        ),
-        vscode.commands.registerCommand(
-            "extension.devicescript.register.edit",
-            async (item: JDomRegisterTreeItem) => {
-                if (!item) return
-
-                const { register } = item
-                const { specification } = register
-
-                if (!specification) {
-                    vscode.window.showWarningMessage(
-                        `DeviceScript: no specification found for register.`
-                    )
-                    return
-                }
-
-                const { kind, fields, description } = specification
-                if (kind !== "rw") {
-                    vscode.window.showWarningMessage(
-                        "DeviceScript: register cannot be modified."
-                    )
-                    return
-                }
-
-                await register.refresh()
-                const { name, unpackedValue } = register
-                const title = `Edit register ${name}`
-                if (fields?.length === 1) {
-                    const field = fields[0]
-                    const { type, unit, isFloat } = field
-                    const prompt = `${description} ${
-                        unit ? `(${prettyUnit(unit)})` : ""
-                    }`
-                    const value = unpackedValue?.[0]
-
-                    // strings
-                    if (type === "string" || type === "string0") {
-                        const v = value as string
-                        const res = await vscode.window.showInputBox({
-                            title,
-                            prompt,
-                            value: v || "",
-                        })
-                        if (res !== undefined && v !== res) {
-                            await register.sendSetStringAsync(res, true)
-                        }
-                        return
-                    }
-                    // boolean
-                    else if (type === "bool") {
-                        const v = value ? "yes" : "no"
-                        const res = await vscode.window.showQuickPick(
-                            ["true", "false"],
-                            <vscode.QuickPickOptions>{
-                                title,
-                            }
-                        )
-                        if (res !== undefined && v !== res) {
-                            await register.sendSetBoolAsync(
-                                res === "true",
-                                true
-                            )
-                        }
-                        return
-                    }
-                    // float
-                    else if (isFloat) {
-                        const v = (value as number)?.toLocaleString()
-                        // TODO: min, max
-                        const res = await vscode.window.showInputBox({
-                            title,
-                            prompt,
-                            value: v || "0",
-                            validateInput: i => {
-                                if (isNaN(parseFloat(i)))
-                                    return "invalid number format"
-                                return undefined
-                            },
-                        })
-                        if (res !== undefined && v !== res) {
-                            const newValue = parseFloat(res)
-                            await register.sendSetPackedAsync([newValue], true)
-                        }
-                        return
-                    }
-                    // int, uint
-                    else if (field.unit) {
-                        const v = (value as number)?.toLocaleString()
-                        // TODO: min, max
-                        const res = await vscode.window.showInputBox({
-                            title,
-                            prompt,
-                            value: v || "0",
-                            validateInput: i => {
-                                if (isNaN(parseInt(i)))
-                                    return `invalid ${field.type} number format`
-                                return undefined
-                            },
-                        })
-                        if (res !== undefined && v !== res) {
-                            const newValue = parseInt(res)
-                            await register.sendSetPackedAsync([newValue], true)
-                        }
-                        return
-                    }
-                }
-
-                vscode.window.showWarningMessage(
-                    "DeviceScript: Sorry, this register cannot be edited by the extension."
-                )
-            }
-        ),
-        vscode.commands.registerCommand(
-            "extension.devicescript.watch.clear",
-            async () => {
-                await extensionState.updateWatches([])
-                jdomWatchTreeDataProvider.refresh()
-            }
-        ),
-        vscode.commands.registerCommand(
-            "extension.devicescript.watch.add",
-            async (item: JDomTreeItem) => {
-                if (!item) return
-                console.log(`Watch ${item.node}`)
-                const id = item.node.id
-                const { watches } = extensionState
-                if (!watches.find(w => w.id === id)) {
-                    const label = item.node.friendlyName
-                    const icon = (item.iconPath as vscode.ThemeIcon)?.id
-                    await extensionState.updateWatches([
-                        ...watches,
-                        <NodeWatch>{ id, label, icon },
-                    ])
-                    item.refresh()
-                    jdomWatchTreeDataProvider.refresh()
-                }
-            }
-        ),
-        vscode.commands.registerCommand(
-            "extension.devicescript.watch.remove",
-            async (item: JDomTreeItem) => {
-                if (!item) return
-                console.log(`Unwatch ${item.node}`)
-                const id = item.node.id
-                const { watches } = extensionState
-                if (watches.find(w => w.id === id)) {
-                    await extensionState.updateWatches(
-                        watches.filter(w => w.id !== id)
-                    )
-                    item.refresh()
-                    jdomWatchTreeDataProvider.refresh()
-                }
-            }
         )
     )
 
+    activateTreeViews(extensionState)
     activateMainStatusBar(extensionState)
 
     // packet trace
